+++ src/agent/calendar_agent/tools/non_sensitive_tools.py
@@ -10,8 +10,9 @@ from typing import Union
 
 from googleapiclient.discovery import build
 from langchain_core.tools import tool
-from tools.get_credentials import get_credentials
-from utils import read_personal_info
+
+from ...shared.get_credentials import get_credentials
+from ...shared.utils import read_personal_info
 
 
 @tool
