+++ src/agent/calendar_agent/tools/sensitive_tools.py
@@ -7,8 +7,9 @@ from typing import Optional
 
 from googleapiclient.discovery import build
 from langchain_core.tools import tool
-from tools.get_credentials import get_credentials
-from utils import read_personal_info
+
+from ...shared.get_credentials import get_credentials
+from ...shared.utils import read_personal_info
 
 
 def validate_datetime(dt_str: str) -> bool:
