+++ src/agent/gmail_agent/tools/non_sensitive_tools.py
@@ -7,9 +7,9 @@ from typing import Optional
 
 from googleapiclient.discovery import build
 from langchain_core.tools import tool
-from tools.get_credentials import get_credentials
-from utils import read_personal_info
 
+from ...shared.get_credentials import get_credentials
+from ...shared.utils import read_personal_info
 from .utils import extract_clean_text
