+++ src/agent/shared/get_credentials.py
@@ -9,7 +9,8 @@ from google.auth.transport.requests import Request
 from google.oauth2.credentials import Credentials
 from google_auth_oauthlib.flow import InstalledAppFlow
 from googleapiclient.discovery import build
-from utils import save_personal_info
+
+from .utils import save_personal_info
 
 load_dotenv()
 logging.basicConfig(
